Rename VMAX driver folder to 'dellemc' (#62)
* Move dellemc_storage to dellemc

* Update for folder name change

* Rename 'dellemc' to 'dell_emc' as in opensource drivers
